streamlit_app/requirements.txt
numpy==1.26.4
pandas==2.2.2
streamlit==1.36.0
Pillow==10.4.0
seaborn==0.13.2
matplotlib==3.9.1
scikit-learn==1.4.2
tqdm==4.66.4
scipy==1.14.0
imbalanced-learn==0.12.3
statsmodels==0.14.2
xgboost==2.1.0
shap==0.46.0